+++ src/Parser/readProperty.ts
@@ -1,20 +1,28 @@
 ﻿import { spParserArgs } from "./spParser";
 import { PropertyItem } from "../Backend/Items/spPropertyItem";
-import { ParsedComment, ParsedID, ParserLocation } from "./interfaces";
+import {
+  MethodDeclaration,
+  MethodmapNativeForwardDeclaration,
+  ParsedComment,
+  ParsedID,
+  ParserLocation,
+} from "./interfaces";
 import { parsedLocToRange } from "./utils";
 import { MethodMapItem } from "../Backend/Items/spMethodmapItem";
-import { EnumStructItem } from "../Backend/Items/spEnumStructItem";
 import { processDocStringComment } from "./processComment";
+import { readFunctionAndMethod } from "./readFunctionAndMethod";
 
 export function readProperty(
   parserArgs: spParserArgs,
   id: ParsedID,
   loc: ParserLocation,
-  parent: MethodMapItem | EnumStructItem,
+  parent: MethodMapItem,
   docstring: ParsedComment,
-  returnType: ParsedID
+  returnType: ParsedID,
+  body: (MethodDeclaration | MethodmapNativeForwardDeclaration)[]
 ): void {
   const range = parsedLocToRange(id.loc, parserArgs);
+  const fullRange = parsedLocToRange(loc, parserArgs);
   const { doc, dep } = processDocStringComment(docstring);
   const propertyItem = new PropertyItem(
     parent,
@@ -23,7 +31,21 @@ export function readProperty(
     id.id,
     doc,
     range,
+    fullRange,
     returnType.id
   );
   parserArgs.fileItems.set(`${id.id}-${parent.name}`, propertyItem);
+  body.forEach((e) => {
+    readFunctionAndMethod(
+      parserArgs,
+      e.accessModifier,
+      e.returnType,
+      e.id,
+      e.loc,
+      e.doc,
+      e.params,
+      e.body,
+      propertyItem
+    );
+  });
 }